refactor(corridors): split segments logic to module and fix type-only imports
- Extract segments.ts (isDashed/extract/buildContinuousTrack*)
- Use type-only imports for LonLatAlt/Segment to avoid runtime bindings
