+++ frontend/map-corridors/src/corridors/preciseCorridor.ts
@@ -1,10 +1,10 @@
 import type { Feature, FeatureCollection, GeoJSON, LineString, Point, Position } from 'geojson'
 import { lineString, length as turfLength, getCoord, bearing as turfBearing, destination, point, nearestPointOnLine } from '@turf/turf'
+import type { LonLatAlt, Segment } from './segments'
+import { calculateDistance, buildContinuousTrackWithSources } from './segments'
 
 const DEBUG = (import.meta as any)?.env?.VITE_DEBUG_CORRIDORS === 'true' || (import.meta as any)?.env?.VITE_DEBUG_CORRIDORS === '1'
 const log = (...args: any[]) => { if (DEBUG) console.log(...args) }
-
-type LonLatAlt = [number, number, number?]
 
 export type CorridorOutput = {
   left: Feature<LineString>
@@ -21,136 +21,7 @@ function projectCoordinate(origin: LonLatAlt, bearingDeg: number, distanceMeters
   return [lon, lat, origin[2]]
 }
 
-function calculateDistance(coord1: LonLatAlt, coord2: LonLatAlt): number {
-  const [lon1, lat1] = coord1
-  const [lon2, lat2] = coord2
-  const R = 6371000 // Earth's radius in meters
-  const lat1Rad = lat1 * Math.PI / 180
-  const lat2Rad = lat2 * Math.PI / 180
-  const deltaLat = (lat2 - lat1) * Math.PI / 180
-  const deltaLon = (lon2 - lon1) * Math.PI / 180
-  const a = Math.sin(deltaLat/2) * Math.sin(deltaLat/2) +
-           Math.cos(lat1Rad) * Math.cos(lat2Rad) *
-           Math.sin(deltaLon/2) * Math.sin(deltaLon/2)
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a))
-  return R * c
-}
-
-function isDashedConnectorLine(coords: LonLatAlt[]): boolean {
-  // Skip segments that are not simple 2-point lines
-  if (coords.length !== 2) {
-    return false
-  }
-  // Calculate segment length
-  const length = calculateDistance(coords[0], coords[1])
-  // Very short segments are likely dashed connectors
-  return length < 500  // Less than 500m (match backend)
-}
-
-type Segment = {
-  index: number
-  coordinates: LonLatAlt[]
-}
-
-function extractAllSegments(input: GeoJSON): Segment[] {
-  const segments: Segment[] = []
-  let index = 0
-  function extract(g: any) {
-    if (!g) return
-    if (g.type === 'FeatureCollection') {
-      for (const f of (g as FeatureCollection).features) extract(f)
-    } else if (g.type === 'Feature') {
-      const f = g as Feature
-      const geom = f.geometry
-      if (geom?.type === 'LineString') {
-        const ls = geom as LineString
-        const coords = ls.coordinates as LonLatAlt[]
-        // Skip turning point markers (3-coordinate segments)
-        if (coords.length === 3) return
-        segments.push({ index: index++, coordinates: coords })
-      }
-    } else if (g.type === 'LineString') {
-      const ls = g as LineString
-      const coords = ls.coordinates as LonLatAlt[]
-      // Skip turning point markers (3-coordinate segments)
-      if (coords.length === 3) return
-      segments.push({ index: index++, coordinates: coords })
-    }
-  }
-  extract(input)
-  return segments
-}
-
-export function buildContinuousTrackWithSources(input: GeoJSON): { track: LonLatAlt[], sourceSegIdx: number[], gapAfterIndex: boolean[], segments: Segment[], mainSegmentIndexSet: Set<number> } {
-  // Extract all segments
-  const allSegments = extractAllSegments(input)
-  
-  // Classify segments: main track vs dashed connectors
-  const mainTrackSegments = allSegments.filter(seg => !isDashedConnectorLine(seg.coordinates))
-  const dashedConnectors = allSegments.filter(seg => isDashedConnectorLine(seg.coordinates))
-  
-  log(`=== PROCESSING SUMMARY ===`)
-  log(`Total line segments found: ${allSegments.length}`)
-  log(`Main track segments: ${mainTrackSegments.length}`)
-  log(`Dashed connectors (excluded): ${dashedConnectors.length}`)
-  
-  // Build detailed continuous track from main segments only
-  if (mainTrackSegments.length === 0) return { track: [], sourceSegIdx: [], gapAfterIndex: [], segments: allSegments, mainSegmentIndexSet: new Set() }
-  
-  // Sort segments by index to maintain original order
-  const sortedSegments = mainTrackSegments.sort((a, b) => a.index - b.index)
-  
-  const detailedTrack: LonLatAlt[] = []
-  const sourceSegIdx: number[] = []
-  const gapAfterIndex: boolean[] = [] // length will be track.length - 1
-  
-  for (let i = 0; i < sortedSegments.length; i++) {
-    const segment = sortedSegments[i]
-    const coords = segment.coordinates
-    
-    if (i === 0) {
-      // First segment: append all points, all edges are contiguous
-      for (let k = 0; k < coords.length; k++) {
-        detailedTrack.push(coords[k])
-        sourceSegIdx.push(segment.index)
-        if (k > 0) gapAfterIndex.push(false)
-      }
-    } else {
-      const lastPoint = detailedTrack[detailedTrack.length - 1]
-      const firstPoint = coords[0]
-      const distance = calculateDistance(lastPoint, firstPoint)
-      if (distance < 50) {
-        // Connected: skip duplicate first, append rest; edges contiguous
-        for (let k = 1; k < coords.length; k++) {
-          detailedTrack.push(coords[k])
-          sourceSegIdx.push(segment.index)
-          gapAfterIndex.push(false)
-        }
-      } else {
-        // Gap exists between previous last and this segment's first point
-        // Append first point and mark the edge as a gap
-        detailedTrack.push(coords[0])
-        sourceSegIdx.push(segment.index)
-        gapAfterIndex.push(true)
-        // Append remaining points with contiguous edges
-        for (let k = 1; k < coords.length; k++) {
-          detailedTrack.push(coords[k])
-          sourceSegIdx.push(segment.index)
-          gapAfterIndex.push(false)
-        }
-      }
-    }
-  }
-  
-  log(`Built detailed track: ${detailedTrack.length} total points from ${mainTrackSegments.length} segments`)
-  
-  const mainSet = new Set<number>(mainTrackSegments.map(s => s.index))
-  return { track: detailedTrack, sourceSegIdx, gapAfterIndex, segments: allSegments, mainSegmentIndexSet: mainSet }
-}
-
-export function buildContinuousTrack(input: GeoJSON): LonLatAlt[] {
-  return buildContinuousTrackWithSources(input).track
-}
+// moved: isDashedConnectorLine/extract/buildContinuousTrack* to segments.ts
 
 export function generateLeftRightCorridor(track: LonLatAlt[], corridorDistanceM = 300): CorridorOutput | null {
   if (track.length < 2) return null
